Migrate middlewares outside the services folder
Services does not really means what this code does.

This is a problem, as it makes the debugging tedious for a
new contributor.

This fixes it by moving the middlewares to their own package.
