+++ internal/middlewares/prometheus.go
@@ -1,4 +1,4 @@
-package utils
+package middlewares
 
 import (
 	"log"
@@ -6,9 +6,16 @@ import (
 
 	"github.com/gorilla/mux"
 	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/promauto"
 )
 
-func PrometheusMiddleware(next http.Handler) http.Handler {
+var Histogram = promauto.NewHistogramVec(prometheus.HistogramOpts{
+	Name:    "kubi_http_requests",
+	Help:    "Time per requests",
+	Buckets: []float64{1, 2, 5, 6, 10}, //defining small buckets as this app should not take more than 1 sec to respond
+}, []string{"path"}) // this will be partitioned by the HTTP code.
+
+func Prometheus(next http.Handler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		route := mux.CurrentRoute(r)
 		path, err := route.GetPathTemplate()
