+++ internal/utils/prometheus.go
@@ -10,12 +10,6 @@ var TokenCounter = promauto.NewCounterVec(prometheus.CounterOpts{
 	Help: "Total number of tokens issued",
 }, []string{"status"})
 
-var Histogram = promauto.NewHistogramVec(prometheus.HistogramOpts{
-	Name:    "kubi_http_requests",
-	Help:    "Time per requests",
-	Buckets: []float64{1, 2, 5, 6, 10}, //defining small buckets as this app should not take more than 1 sec to respond
-}, []string{"path"}) // this will be partitioned by the HTTP code.
-
 var ProjectCreation = promauto.NewCounterVec(prometheus.CounterOpts{
 	Name: "kubi_project_creation",
 	Help: "Number of project created",
